LPS-53392 Create an specific schemaVersion for this upgrade since the database schema have changed and it's a safe and simple process. We also rename the package since we use the final version
